fix: limit ttl refresh to media storage
