+++ media.pollinations.ai/vitest.config.ts
@@ -1,14 +1,6 @@
-import { fileURLToPath } from "node:url";
 import { defineWorkersConfig } from "@cloudflare/vitest-pool-workers/config";
 
-const sharedSrc = fileURLToPath(new URL("../shared/", import.meta.url));
-
 export default defineWorkersConfig({
-    resolve: {
-        alias: {
-            "@shared": sharedSrc,
-        },
-    },
     test: {
         poolOptions: {
             workers: {
